Small change but Read
Fixed Download button breaking with Expo build for android.

Previous submits where just before build because if it was submitted or not made a difference with the BUILD for some reason did not always acknowledge changes made to the app.json file.

Additionally use
#CMD       expo prebuild
This prebuilds the application for IOS and Android if applicable for both and dose so settings that makes the download button work again as it should and not break on build to android apk file